refactor: 💡 remove route-action usage for groups
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-12605
