+++ ui/admin/app/components/groups/group/actions/index.hbs
@@ -18,7 +18,7 @@
     <dropdown.button
       @style='danger'
       @disabled={{@model.canSave}}
-      {{on 'click' (route-action 'delete' @model)}}
+      {{on 'click' @delete}}
     >
       {{t 'resources.group.actions.delete'}}
     </dropdown.button>
