+++ ui/admin/app/controllers/scopes/scope/groups/group/add-members.js
@@ -4,9 +4,15 @@
  */
 
 import Controller from '@ember/controller';
+import { inject as service } from '@ember/service';
+import { action } from '@ember/object';
+import { loading } from 'ember-loading';
+import { notifySuccess, notifyError } from 'core/decorators/notify';
 
 export default class ScopesScopeGroupsGroupAddMembersController extends Controller {
   // =services
+
+  @service router;
 
   // =attributes
 
@@ -25,5 +31,30 @@ export default class ScopesScopeGroupsGroupAddMembersController extends Controll
       ]);
     const sorted = [global, ...orgsAndProjects].flat();
     return sorted;
+  }
+
+  // =actions
+
+  /**
+   * Adds members to the group and saves, replaces with the members index
+   * route, and notifies the user of success or error.
+   * @param {GroupModel} group
+   * @param {[string]} userIDs
+   */
+  @action
+  @loading
+  @notifyError(({ message }) => message, { catch: true })
+  @notifySuccess('notifications.add-success')
+  async addMembers(group, userIDs) {
+    await group.addMembers(userIDs);
+    await this.router.replaceWith('scopes.scope.groups.group.members');
+  }
+
+  /**
+   * Redirect to group members as if nothing ever happened.
+   */
+  @action
+  async cancel() {
+    await this.router.replaceWith('scopes.scope.groups.group.members');
   }
 }
