+++ ui/admin/app/controllers/scopes/scope/groups/group/index.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeGroupsGroupIndexController extends Controller {
+  @controller('scopes/scope/groups/index') groups;
+}
