+++ ui/admin/app/controllers/scopes/scope/groups/new.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeGroupsNewController extends Controller {
+  @controller('scopes/scope/groups/index') groups;
+}
