+++ ui/admin/app/routes/scopes/scope/groups/group/members.js
@@ -4,12 +4,8 @@
  */
 
 import Route from '@ember/routing/route';
-import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { hash } from 'rsvp';
-import { loading } from 'ember-loading';
-import { confirm } from 'core/decorators/confirm';
-import { notifySuccess, notifyError } from 'core/decorators/notify';
 
 export default class ScopesScopeGroupsGroupMembersRoute extends Route {
   // =services
@@ -36,22 +32,5 @@ export default class ScopesScopeGroupsGroupMembersRoute extends Route {
           )
         : [],
     });
-  }
-
-  // =actions
-
-  /**
-   * Remove a member from the current role and redirect to members index.
-   * @param {GroupModel} group
-   * @param {UserModel} member
-   */
-  @action
-  @loading
-  @confirm('questions.remove-confirm')
-  @notifyError(({ message }) => message, { catch: true })
-  @notifySuccess('notifications.remove-success')
-  async removeMember(group, member) {
-    await group.removeMember(member.id);
-    this.refresh();
   }
 }
