+++ ui/admin/app/templates/scopes/scope/groups/group/add-members.hbs
@@ -86,8 +86,8 @@
     <Form::Group::AddMembers
       @model={{@model.group}}
       @users={{@model.users}}
-      @submit={{route-action 'addMembers' @model.group}}
-      @cancel={{route-action 'cancel'}}
+      @submit={{fn this.addMembers @model.group}}
+      @cancel={{this.cancel}}
     />
   </page.body>
