+++ ui/admin/tests/unit/controllers/scopes/scope/auth-methods/auth-method/accounts/index-test.js
@@ -1,5 +1,5 @@
 import { module, test } from 'qunit';
-import { setupTest } from 'admin/tests/helpers';
+import { setupTest } from 'ember-qunit';
 
 module(
   'Unit | Controller | scopes/scope/auth-methods/auth-method/accounts/index',
