+++ ui/admin/tests/unit/controllers/scopes/scope/groups/group/index-test.js
@@ -1,0 +1,13 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'ember-qunit';
+
+module('Unit | Controller | scopes/scope/groups/group/index', function (hooks) {
+  setupTest(hooks);
+
+  test('it exists', function (assert) {
+    let controller = this.owner.lookup(
+      'controller:scopes/scope/groups/group/index',
+    );
+    assert.ok(controller);
+  });
+});
