+++ ui/admin/tests/unit/controllers/scopes/scope/groups/group/members-test.js
@@ -2,13 +2,13 @@ import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 
 module(
-  'Unit | Controller | scopes/scope/auth-methods/auth-method/accounts/new',
+  'Unit | Controller | scopes/scope/groups/group/members',
   function (hooks) {
     setupTest(hooks);
 
     test('it exists', function (assert) {
       let controller = this.owner.lookup(
-        'controller:scopes/scope/auth-methods/auth-method/accounts/new',
+        'controller:scopes/scope/groups/group/members',
       );
       assert.ok(controller);
     });
